Fix StopAll() in capturer.Storage
There was a race condition in Storage. Event handler routine might exit
before all capturer routines send their exit events. This caused either
a hang or error because of sending over a closed channel.

Additionally stop cmd now logs error when no captureres are running.

Storage pointer handling is reworked. Instad of creating it on start and
setting it to nil on stop, both commands check if the storage is empty.