Fix a deadlock in MultiOutput
Wireshark and File Outputers were both added to MultiOutput's WaitGroup,
however File never sends event over the channel, because it is not an
external process and never dies. This caused the WaitGroup to block the
whole application.

Additionally AddMember() in MultiOutput is renamed to AddExtMember() in
order to make an explicit difference between the Outputers passed on the
construction of MultiOutput and those Outputers added later via the
method.