Build tcpdump filter dynamically
Up to now, the tcpdump filter was 'not port 22'. This patch implements
dynamic filter generation in the format 'not (host SSH_TARGET_IP and
port SSH_TARGET_PORT). Port number is obtained from the configuration,
while the IP address is obtained when the SSH connection is established.
The purpose is to be able to handle domain names as target address.