Add Filter port option in configuration
The target might be behind NAT or some sort of port forwarding may be
configured. As a result the SSH client (rpcap) might consider that the
SSH service runs on port X, while it is redirected to port Y. In this
case rpcap will install wrong port as capture filter for tcpdump and the
user will see rpcap's SSH session (which is undesirable) or even worse -
miss legitimate traffic.

This patch adds new configuration option "Filter port". When set, it is
used in the tcpdump capture filter and overrides the value in "Port",
but only for the tcpdump filter.